gain tax
this program calculates the taxes you have to pay yearly according to Egyptian law. 
this calculates for both business owners and employees, the business owners input their yearly profit and employees input their monthly salary.   
so the law itself is really hard to explain, basically divided into different economical slices based on how much you gain yearly (the slices are the if conditions).
there are percentages that we multiply in (i have put them in an array)
so for example, if you get paid 3000 pounds monthly, we multiply that by 12 to get 36000 as the yearly salary 
and we see which economical slice you are, in this case, you are in the first one
then we subtract 15000 of your yearly salary as tax exemption, then we take the another 15000 and multiply it by 2.5%, and then the 6000 remaining we multiply it by 10% and so on and so forth.